requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
sqlalchemy==2.0.23